refactor(userRegisterTest): 회원가입 화면 위젯 분리
회원가입 화면의 각 UI 구성요소들을 각각의 파일로 분리
